Remove all DOM manipulation from scripts.js and add letter spacing to make recipes more readable
