+++ src/styles.css
@@ -1,6 +1,7 @@
 * {
   background-color: #E7F1F1; 
   margin: 10px;
+  letter-spacing: 2px;
 }
 
 .container {
